Adding the Beacons module to the project
This commit adds the Beacons module to the project and adds a
simple event handler for the event that the module uses to notify
the app that the module is ready.

Now when you run the app, the label that was included with
the default project will display "Beacons Module is ready!" when
the module is ready.