Adding UI elements
This commit adds all the UI elements of this simple app: a number of labels,
a button that will toggle scanning for beacons, and a table that will display
the discovered beacons.

The biggest part of this commit is a widget called 'display'. In this app, we
use this widget to display status information.