Adding region monitoring
This commit adds the primary functionality of the app, monitoring a region.

A region is identified by a UUID, which in this case we store in the REGION_UUID
variable. Monitoring a region means that we scan for beacons that belong to
this region and notify the app whether the user is inside or outside the region.

The startRegionMonitoring and stopRegionMonitoring functions are used
to start and stop the scanning, and the regionStateUpdated event is used
to signal changes in the region's state (inside or outside).