Ranging for beacons
This commit adds the secondary functionality, ranging for beacons.

When the app gets the notification that the user entered a region,
it starts ranging for beacons in that region using the startRangingBeacons
function. This results in periodic rangedBeacons events, roughly one
every second. The app uses these events to determine which beacons
are within range and displays their major number, minor number, and
RSSI value in the table.